requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.1
attrs==26.1.0
bitarray==3.8.0
certifi==2026.2.25
charset-normalizer==3.4.6
ckzg==2.1.7
click==8.3.1
cytoolz==1.1.0
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.8.0
h11==0.16.0
hexbytes==1.3.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
multidict==6.7.1
packaging==26.0
parsimonious==0.10.0
pillow==12.1.1
propcache==0.4.1
psycopg2-binary==2.9.11
py-solc-x==2.0.5
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.2
pyunormalize==17.0.0
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
rlp==4.1.0
rpds-py==0.30.0
sqlparse==0.5.5
toolz==1.1.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
web3==7.14.1
websockets==15.0.1
yarl==1.23.0
